+++ RELEASE_NOTES.md
@@ -23,6 +23,11 @@ New Features
 
 * `ocean.db.tokyocabinet.TokyoCabinetM`
 
-  A new `get()` method has been added which allows a record to be got via a
-  delegate, rather than copying into a buffer. This can be useful in situations
-  where the user doesn't need to store the value.
+  * A new `get()` method has been added which allows a record to be got via a
+    delegate, rather than copying into a buffer. This can be useful in situations
+    where the user doesn't need to store the value.
+
+  * A new non-interruptible `opApply` iterator has been added which implements a
+    much more efficient form of iteration over the in-memory database. This can,
+    obviously, only be used in situations where you are sure that only a single
+    iteration is underway at any one time.